requirements.txt
Django==3.0
psycopg2>=2.7,<3.0
djangorestframework>=3.10
django-cors-headers==3.2.1
djangorestframework-simplejwt==4.4.0
django-extensions==2.2.9

# TODO move to a dev requirements file
#pytest==5.4.2
pytest-django==3.9.0
factory_boy==2.12.0
isort==4.3.21
black==19.10b0
